chore(ui): rename build-crendentials to build-credentials
