+++ ui/actions/providers/providers.ts
@@ -4,7 +4,7 @@ import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 
 import { apiBaseUrl, getAuthHeaders, getFormValue, wait } from "@/lib";
-import { buildSecretConfig } from "@/lib/provider-credentials/build-crendentials";
+import { buildSecretConfig } from "@/lib/provider-credentials/build-credentials";
 import { ProviderCredentialFields } from "@/lib/provider-credentials/provider-credential-fields";
 import { appendSanitizedProviderInFilters } from "@/lib/provider-filters";
 import { handleApiError, handleApiResponse } from "@/lib/server-actions-helper";
